move SimplePostTool into sub dir
