+++ solr/bin/post
@@ -165,7 +165,7 @@ while [ $# -gt 0 ]; do
           if [[ "${1:2:4}" == "url=" ]]; then
             SOLR_URL=${1:6}
           fi
-        else          
+        else
           key="${1:1}"
           shift
   #       echo "$1: PROP"
@@ -231,8 +231,8 @@ if [[ -n "$RECURSIVE" ]]; then
   PROPS+=('-Drecursive=yes')
 fi
 
-echo "$JAVA" -classpath "${TOOL_JAR[0]}" "${PROPS[@]}" org.apache.solr.util.SimplePostTool "${PARAMS[@]}"
-"$JAVA" -classpath "${TOOL_JAR[0]}" "${PROPS[@]}" org.apache.solr.util.SimplePostTool "${PARAMS[@]}"
+echo "$JAVA" -classpath "${TOOL_JAR[0]}" "${PROPS[@]}" org.apache.solr.util.cli.SimplePostTool "${PARAMS[@]}"
+"$JAVA" -classpath "${TOOL_JAR[0]}" "${PROPS[@]}" org.apache.solr.util.cli.SimplePostTool "${PARAMS[@]}"
 
 # post smoker:
 # bin/post -c signals -out yes -type application/json -d '[{"id": 2, "val": 0.47}]'
