+++ solr/core/src/test/org/apache/solr/core/TestSolrConfigHandler.java
@@ -52,8 +52,8 @@ import org.apache.solr.search.SolrCache;
 import org.apache.solr.util.RESTfulServerProvider;
 import org.apache.solr.util.RestTestBase;
 import org.apache.solr.util.RestTestHarness;
-import org.apache.solr.util.SimplePostTool;
 import org.apache.solr.util.TimeOut;
+import org.apache.solr.util.cli.SimplePostTool;
 import org.eclipse.jetty.servlet.ServletHolder;
 import org.junit.After;
 import org.junit.Before;
