+++ solr/core/src/test/org/apache/solr/pkg/TestPackages.java
@@ -77,7 +77,7 @@ import org.apache.solr.search.QParser;
 import org.apache.solr.search.QParserPlugin;
 import org.apache.solr.security.AuthorizationContext;
 import org.apache.solr.util.LogLevel;
-import org.apache.solr.util.SimplePostTool;
+import org.apache.solr.util.cli.SimplePostTool;
 import org.apache.solr.util.plugin.SolrCoreAware;
 import org.apache.zookeeper.data.Stat;
 import org.junit.After;
